Add port in use test for ExternalAPIServer
